requirements.txt
acres==0.5.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
alembic==1.16.5
annotated-types==0.7.0
anyio==4.10.0
asyncer==0.0.8
asyncpg==0.30.0
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.14.2
Brotli==1.1.0
build==1.3.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
chromadb==1.0.20
ci-info==0.3.0
click==8.2.1
cloudpickle==3.1.1
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.9.0
configobj==5.0.9
configparser==7.2.0
ddgs==9.6.1
diskcache==5.6.3
distro==1.9.0
dspy==3.0.3
dspy-ai==3.0.3
duckduckgo_search==8.1.1
durationpy==0.10
et_xmlfile==2.0.0
etelemetry==0.3.1
fastapi==0.116.1
fastuuid==0.13.5
filelock==3.19.1
flatbuffers==25.2.10
frontend==0.0.3
frozenlist==1.8.0
fsspec==2025.7.0
gepa==0.0.7
google-auth==2.40.3
googleapis-common-protos==1.70.0
greenlet==3.2.4
groq==0.31.0
grpcio==1.74.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.30.0
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.34.4
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.0
joblib==1.5.2
json_repair==0.52.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
kubernetes==33.1.0
langchain-core==0.3.75
langchain-text-splitters==0.3.11
langsmith==0.4.21
litellm==1.78.2
looseversion==1.3.0
lxml==6.0.1
magicattr==0.1.6
Mako==1.3.10
markdown-it-py==4.0.0
markdownify==1.2.0
MarkupSafe==3.0.2
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
multidict==6.7.0
networkx==3.5
nibabel==5.3.2
nipype==1.10.0
nltk==3.9.1
numpy==2.2.6
oauthlib==3.3.1
onnxruntime==1.22.1
openai==2.3.0
opencv-python==4.12.0.88
openpyxl==3.1.5
opentelemetry-api==1.36.0
opentelemetry-exporter-otlp-proto-common==1.36.0
opentelemetry-exporter-otlp-proto-grpc==1.36.0
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
optuna==4.5.0
orjson==3.11.3
overrides==7.7.0
packaging==25.0
pandas==2.3.2
pathlib==1.0.1
pillow==11.3.0
posthog==5.4.0
primp==0.15.0
propcache==0.4.1
protobuf==6.32.0
prov==2.1.1
psycopg2==2.9.10
psycopg2-binary==2.9.10
puremagic==1.30
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pydantic==2.12.2
pydantic_core==2.41.4
pydot==4.0.1
Pygments==2.19.2
PyMuPDF==1.26.4
pyparsing==3.2.3
PyPDF2==3.0.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-magic-bin==0.4.14
python-multipart==0.0.20
pytz==2025.2
pyxnat==1.6.3
PyYAML==6.0.2
rdflib==7.1.4
referencing==0.36.2
regex==2025.8.29
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.1.0
rpds-py==0.27.1
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.16.1
sentence-transformers==5.1.0
sentencepiece==0.2.1
setuptools==80.9.0
shellingham==1.5.4
simplejson==3.20.1
six==1.17.0
smolagents==1.22.0
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.8
SQLAlchemy==2.0.43
starlette==0.47.3
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.0
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
traits==7.0.2
transformers==4.56.0
typer==0.17.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
unicorn==2.1.3
urllib3==2.5.0
uvicorn==0.35.0
watchfiles==1.1.0
websocket-client==1.8.0
websockets==15.0.1
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.24.0
